% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpMatrix.R
\docType{data}
\name{mlpMatrix}
\alias{mlpMatrix}
\title{Simulated data used to demonstrate the Mlp method.}
\format{Matrix
\describe{
\item{rownames}{Groups}
\item{colnames}{dimension number}
}}
\usage{
mlpMatrix
}
\value{
simulated matrix
}
\description{
Mlp demonstration matrix.
}
\examples{
mlpMatrix

}
