% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpMatrix.R
\docType{data}
\name{pcpMatrix}
\alias{pcpMatrix}
\title{Simulated data used to demonstrate the Pcp method.}
\format{Matrix
\describe{
\item{rownames}{Groups}
\item{colnames}{dimension number}
}}
\usage{
pcpMatrix
}
\value{
simulated matrix
}
\description{
Pcp demonstration matrix.
}
\examples{
pcpMatrix

}
