% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSignificance-package.R
\docType{package}
\name{ClusterSignificance-package}
\alias{ClusterSignificance-package}
\alias{ClusterSignificance}
\title{The ClusterSignificance package provides tools to assess
if clusters have a separation different from random or permuted
data.}
\description{
The ClusterSignificance package provides tools to assess
if clusters have a separation different from random or permuted
data. ClusterSignificance investigates clusters of two or more
groups by first, projecting all points onto a one dimensional
line. Cluster separations are then scored and the probability
of the seen separation being due to chance is evaluated using a
permutation method.
}
\details{
\tabular{ll}{ Package: \tab ClusterSignificance\cr Type: \tab Package\cr
Version: \tab 1.0\cr Date: \tab 2016-02-28\cr License: \tab GPL-3\cr }
}
\references{
Reference to published application note (work in progress)
}
\author{
Author: Jason T. Serviss, Jesper R. Gadin
}
\keyword{package}

